refactor(luasnip): simplify structure
